+++ CHANGES.md
@@ -6,7 +6,7 @@
 - Partial support for constrained pointer fields [#58](https://github.com/OCamlPro/seacoral/pull/58)
 - Argument `--initialization` to `seacoral check` to assess proper project initialization after configuration checks [#73](https://github.com/OCamlPro/seacoral/pull/73)
 - New subcommand for replaying tests only [#47](https://github.com/OCamlPro/seacoral/pull/47)
-- Association between labels and tests in the corpus [#53](https://github.com/OCamlPro/seacoral/pull/53)
+- Association between labels and tests in the corpus [#53](https://github.com/OCamlPro/seacoral/pull/53) [#93](https://github.com/OCamlPro/seacoral/pull/93)
 - A pretty error message when entrypoints or cover targers are missing from the C code [#52](https://github.com/OCamlPro/seacoral/pull/52)
 - Reporting of RTEs and unsatisfied oracles by CBMC in `assert` and `clabel` modes [#48](https://github.com/ocamlpro/seacoral/pull/48)
 - Comment with test outcomes in generated testsuites [#44](https://github.com/ocamlpro/seacoral/pull/44)